+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Different games running under the bot can now have messages passed between them
 - Stopping a game that is not running now results in a message stating that the game cannot be stopped
+- restart command (will not work in Delve)
 
 ### Fixed
 - unset formatters for GameRegexps and various other configs no longer error on start or in use.